Move helper-function, add votes
